Code cleanup

+++ code/examples/database/all_drivers_test.cpp
@@ -218,7 +218,7 @@ int testDatabase(const string& connectionString)
 
         // Here is the example of using parameters by index.
         // This is the even faster than stream
-        insertRecordQuery.param(uint32_t(0)) = 3;
+        insertRecordQuery.param(size_t(0)) = 3;
         insertRecordQuery.param(1) = "UTF-8: тестик (Russian, 6 chars)";
         insertRecordQuery.param(2).setNull(VAR_STRING);
         insertRecordQuery.param(3).setDateTime(DateTime::Now(), true);
